split integration and step tests
